backend/requirements.txt
# Core Django
Django>=4.2.0
djangorestframework>=3.14.0
django-cors-headers>=4.0.0

# Authentication
djangorestframework-simplejwt>=5.2.2

# Database
psycopg2-binary>=2.9.5  # For PostgreSQL

# API Documentation
drf-yasg>=1.21.5

# Filtering
django-filter>=22.1

# Environment variables
python-dotenv>=1.0.0

# Image processing
Pillow>=9.5.0

# Async
channels>=4.0.0  # For WebSocket support if needed

# Testing
pytest>=7.2.0
pytest-django>=4.5.2
pytest-cov>=4.0.0

# Code quality
black>=23.1.0
isort>=5.12.0
flake8>=6.0.0

# Development
ipython>=8.10.0
django-extensions>=3.2.3

# Deployment
gunicorn>=20.1.0
whitenoise>=6.4.0
